actionlint.rs alphavet.rs deadcode.rs depth.rs dprint.rs gh.rs ghokin.rs go.rs gofumpt.rs golangci_lint.rs goreleaser.rs mod.rs nodejs.rs npm.rs scc.rs shellcheck.rs shfmt.rs args.rs command.rs versions.rs github_releases.rs github_tags.rs

+++ src/apps/gofumpt.rs
@@ -53,7 +53,7 @@ impl App for Gofumpt {
         yard.load_app(self.name(), version, self.executable_filename(platform))
     }
 
-    fn versions(&self, amount: u8, output: &dyn Output) -> Result<Vec<String>> {
+    fn versions(&self, amount: usize, output: &dyn Output) -> Result<Vec<String>> {
         github_releases::versions(ORG, REPO, amount, output)
     }
 }
